+++ src/main/scala/seqOps.scala
@@ -28,31 +28,6 @@ case object seqOps {
     }
   }
 
-  // implicit class ArrayOps[A](val arr: Seq[A]) extends AnyVal {
-  //
-  //   import collection.mutable.ArrayBuffer
-  //
-  //   def splitSegments(pred: A => Boolean)(implicit ct: reflect.ClassTag[A]): Seq[Seq[A]] =
-  //     splitSegments_rec(arr, new ArrayBuffer[Seq[A]](), pred)
-  //
-  //   @annotation.tailrec
-  //   private def splitSegments_rec[X: reflect.ClassTag](xs: Seq[X], acc: ArrayBuffer[Seq[X]], pred: X => Boolean): Seq[Seq[X]] = {
-  //
-  //     val (segment, rest) = xs span { x => (!pred(x)) }
-  //
-  //     val nextAcc = if(segment.isEmpty) acc else acc :+ segment
-  //
-  //     if(rest.isEmpty)
-  //       nextAcc.toArray
-  //     else
-  //       splitSegments_rec(
-  //         xs    = rest dropWhile { x => pred(x) },
-  //         acc   = nextAcc,
-  //         pred  = pred
-  //       )
-  //   }
-  // }
-
   // String is not Seq[Char], so need to repeat all this
   implicit class StringOps(val str: String) extends AnyVal {
